A bunch of queue work

+++ Shoko.Server/Databases/NHIbernate/DateOnlyConverter.cs
@@ -1,3 +1,4 @@
+#nullable enable
 using System;
 using System.Collections;
 using System.ComponentModel;
@@ -10,7 +11,6 @@ using NHibernate.SqlTypes;
 using NHibernate.UserTypes;
 using Shoko.Server.Extensions;
 
-#nullable enable
 namespace Shoko.Server.Databases.NHibernate;
 
 public class DateOnlyConverter : TypeConverter, IUserType
@@ -58,7 +58,7 @@ public class DateOnlyConverter : TypeConverter, IUserType
             },
             "System.Int64" => value switch
             {
-                DateOnly i => (long)i.ToDateTime(TimeOnly.MinValue).Ticks,
+                DateOnly i => i.ToDateTime(TimeOnly.MinValue).Ticks,
                 _ => null,
             },
             "System.String" => value switch
